[mod/wiki] Default for groupid on new pages to 0
Fix MDL-69789

Fixes the issue, that when creating a new wiki page via webservices and
not providing a groupid, the creating of the page in the database fails
with the error message, that the group id collumn can not be null. It
has to be 0 instread.

For a full report see https://tracker.moodle.org/browse/MDL-69789